+++ dev/examples/staggerChildren.tsx
@@ -26,17 +26,32 @@ const items = [0, 1, 2, 3, 4, 5]
 
 export const App = () => {
     const sidebarPoses = {
-        open: [{ x: 0 }, { beforeChildren: true, staggerChildren: 75 }],
+        open: [{ x: 0 }, { beforeChildren: true, staggerChildren: 0.075 }],
         closed: { x: -180 },
     }
 
     const itemPoses = {
-        open: [{ scale: 1, opacity: 1 }, { scale: { type: "spring", stiffness: 400, velocity: 40, damping: 20 } }],
+        open: [
+            { scale: 1, opacity: 1 },
+            {
+                scale: {
+                    type: "spring",
+                    stiffness: 400,
+                    velocity: 40,
+                    damping: 20,
+                },
+            },
+        ],
         closed: [{ scale: 0.5, opacity: 0 }],
     }
 
     return (
-        <motion.ul animate={sidebarPoses} initialPose="closed" pose="open" style={sidebarStyle}>
+        <motion.ul
+            animate={sidebarPoses}
+            initialPose="closed"
+            pose="open"
+            style={sidebarStyle}
+        >
             {items.map(i => (
                 <motion.li animate={itemPoses} inherit style={itemStyle} />
             ))}